# openapi.yml
openapi: 3.1.0
info:
  title: TechCo Backend API
  description: API documentation for the TechCo backend services.
  version: 1.0.0
servers:
  - url: https://api.techco.com/v1
    description: Production server
  - url: https://staging-api.techco.com/v1
    description: Staging server
paths:
...
  /hiring_process/assistant/response:
    post:
      summary: Get response from assistant
      description: Get response from assistant based on the provided request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantResponseRequest"
      responses:
        "200":
          description: Assistant response retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Exception response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exception"
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
    Exception:
      type: object
      properties:
        error:
          type: string
          example: Unexpected error
...
    AssistantResponseRequest:
      type: object
      properties:
        hiring_process_id:
          type: string
          example: "6777862efe77308365432a36"
        run_id:
          type: string
          example: "6777862efe77308365432a36"
        thread_id:
          type: string
          example: "6777862efe77308365432a36"
        assistan_type:
          type: string
          enum:
            - position_assistant
            - technical_assessment_assistant
            - soft_assessment_assistant
      required:
        - hiring_process_id
        - run_id
        - thread_id
        - assistan_type